+++ plugin/src/test/kotlin/org/jlleitschuh/gradle/ktlint/DisabledRulesTest.kt
@@ -1,11 +1,9 @@
 package org.jlleitschuh.gradle.ktlint
 
-import net.swiftzer.semver.SemVer
 import org.assertj.core.api.Assertions.assertThat
 import org.gradle.api.logging.Logging
 import org.gradle.testkit.runner.TaskOutcome
 import org.gradle.util.GradleVersion
-import org.jlleitschuh.gradle.ktlint.testdsl.CommonTest
 import org.jlleitschuh.gradle.ktlint.testdsl.GradleTestVersions
 import org.jlleitschuh.gradle.ktlint.testdsl.build
 import org.jlleitschuh.gradle.ktlint.testdsl.project
@@ -17,76 +15,25 @@ import org.junit.jupiter.params.provider.ArgumentsSource
 class DisabledRulesTest : AbstractPluginTest() {
     private val logger = Logging.getLogger(DisabledRulesTest::class.java)
 
-    @DisplayName("Should lint without errors when 'final-newline' rule is disabled")
-    @ParameterizedTest(name = "{0} with KtLint {1}: {displayName}")
-    @ArgumentsSource(KtLintSupportedVersionsTest.SupportedKtlintVersionsProvider::class)
-    fun lintDisabledRuleFinalNewline(gradleVersion: GradleVersion, ktLintVersion: String) {
-        if (SemVer.parse(ktLintVersion) < SemVer.parse("0.49.0")) {
-            project(gradleVersion) {
-                //language=Groovy
-                buildGradle.appendText(
-                    """
-                    ktlint {
-                        version = "$ktLintVersion"
-                        disabledRules = ["final-newline"]
-                    }
-                    """.trimIndent()
-                )
-
-                createSourceFile(
-                    "src/main/kotlin/CleanSource.kt",
-                    "val foo = \"bar\""
-                )
-
-                if (SemVer.parse(ktLintVersion) < SemVer.parse("0.48.0")) {
-                    build(CHECK_PARENT_TASK_NAME) {
-                        assertThat(task(":$mainSourceSetCheckTaskName")?.outcome).isEqualTo(TaskOutcome.SUCCESS)
-                        assertThat(output).doesNotContain("Property 'ktlint_disabled_rules' is deprecated")
-                    }
-                } else {
-                    build(CHECK_PARENT_TASK_NAME) {
-                        assertThat(task(":$mainSourceSetCheckTaskName")?.outcome).isEqualTo(TaskOutcome.SUCCESS)
-                        assertThat(output)
-                            .`as`("old disabled_rules list is deprecated in 0.48, slated for removal in 0.49")
-                            .contains("Property 'ktlint_disabled_rules' is deprecated")
-                    }
-                }
-            }
-        } else {
-            logger.info("ktlint $ktLintVersion only supports disabling rules via .editorconfig")
-        }
-    }
-
     @DisplayName("Should lint without errors when 'final-newline' rule is disabled via editorconfig")
     @ParameterizedTest(name = "{0} with KtLint {1}: {displayName}")
     @ArgumentsSource(KtLintSupportedVersionsTest.SupportedKtlintVersionsProvider::class)
     fun lintDisabledRuleFinalNewlineEditorconfig(gradleVersion: GradleVersion, ktLintVersion: String) {
         project(gradleVersion) {
-            if (SemVer.parse(ktLintVersion) >= SemVer.parse("0.48.0")) {
-                // new way of disabling rules introduced in 0.48
-                editorConfig.appendText(
-                    """
-                        root = true
+            editorConfig.appendText(
+                """
+root = true
 
-                        [*.kt]
-                        ktlint_standard_final-newline = disabled
-                    """.trimIndent()
-                )
-            } else {
-                // disabled rules property renamed in 0.47 (disabled_rules is still honored, but will print a warning)
-                editorConfig.appendText(
-                    """
-                        root = true
-                        [*.kt]
-                        ktlint_disabled_rules = final-newline
-                    """.trimIndent()
-                )
-            }
+[*.kt]
+ktlint_standard_final-newline = disabled
+"""
+            )
+
             //language=Groovy
             buildGradle.appendText(
                 """
-                    ktlint.version = "$ktLintVersion"
-                """.trimIndent()
+ktlint.version = "$ktLintVersion"
+"""
             )
 
             createSourceFile(
@@ -98,68 +45,6 @@ class DisabledRulesTest : AbstractPluginTest() {
                 assertThat(task(":$mainSourceSetCheckTaskName")?.outcome).isEqualTo(TaskOutcome.SUCCESS)
                 assertThat(output).doesNotContain("Property 'ktlint_disabled_rules' is deprecated")
                 assertThat(output).doesNotContain("Property 'disabled_rules' is deprecated")
-            }
-        }
-    }
-
-    @DisplayName("Should lint without errors when 'final-newline' and 'no-consecutive-blank-lines' are disabled")
-    @CommonTest
-    fun lintOnMultipleDisabledRules(gradleVersion: GradleVersion) {
-        project(gradleVersion) {
-            //language=Groovy
-            buildGradle.appendText(
-                """
-                ktlint.version = "0.47.1"
-                ktlint.disabledRules = ["final-newline", "no-consecutive-blank-lines"]
-                """.trimIndent()
-            )
-
-            createSourceFile(
-                "src/main/kotlin/CleanSource.kt",
-                """
-                fun some() {
-                    print("Woohoo!")
-                }
-
-
-                val foo = "bar"
-                """.trimIndent()
-            )
-
-            build(CHECK_PARENT_TASK_NAME) {
-                assertThat(task(":$mainSourceSetCheckTaskName")?.outcome).isEqualTo(TaskOutcome.SUCCESS)
-            }
-        }
-    }
-
-    @DisplayName("Should lint without errors when 'no-consecutive-blank-lines' are disabled in the code")
-    @CommonTest
-    fun lintRuleDisabledInTheCode(gradleVersion: GradleVersion) {
-        project(gradleVersion) {
-            //language=Groovy
-            buildGradle.appendText(
-                """
-                ktlint.version = "0.47.1"
-                """.trimIndent()
-            )
-            createSourceFile(
-                "src/main/kotlin/CleanSource.kt",
-                """
-                /* ktlint-disable no-consecutive-blank-lines */
-                fun some() {
-                    print("Woohoo!")
-                }
-
-
-                /* ktlint-enable no-consecutive-blank-lines */
-
-                val foo = "bar"
-
-                """.trimIndent()
-            )
-
-            build(CHECK_PARENT_TASK_NAME) {
-                assertThat(task(":$mainSourceSetCheckTaskName")?.outcome).isEqualTo(TaskOutcome.SUCCESS)
             }
         }
     }