+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+## [13.0.0] - 2025-07-07
+
+- remove support for ktlint &lt; 1 [#889](https://github.com/JLLeitschuh/ktlint-gradle/pull/889)
+- increase minimum kotlin version to 1.7 [#889](https://github.com/JLLeitschuh/ktlint-gradle/pull/889)
+- update default ktlint version to 1.5 [#889](https://github.com/JLLeitschuh/ktlint-gradle/pull/889)
+
+## [12.3.0] - 2025-05-22
+
 ### Changed
 
 - Updated build to use Gradle version 8.13 [#850](https://github.com/JLLeitschuh/ktlint-gradle/pull/850)
@@ -770,7 +778,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - Renamed task with name `ktlint` to `ktlintCheck` (#3)
 - Renamed tasks with names `ktlint[source set name]` to `ktlint[source set name]Check` (#3)
 
-[unreleased]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v12.2.0...HEAD
+[unreleased]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v13.0.0...HEAD
+[13.0.0]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v12.3.0...v13.0.0
+[12.3.0]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v12.3.0-rc.2...v12.3.0
+[12.3.0-rc.2]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v12.3.0-rc.1...v12.3.0-rc.2
+[12.3.0-rc.1]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v12.3.0-rc...v12.3.0-rc.1
+[12.3.0-rc]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v12.2.0...v12.3.0-rc
 [12.2.0]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v12.1.2...v12.2.0
 [12.1.2]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v12.1.1...v12.1.2
 [12.1.1]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v12.1.0...v12.1.1